feat: sliderwrapper added to improve performance
